Add support for module imports
