+++ crates/nu-protocol/src/ast/mod.rs
@@ -3,6 +3,7 @@ mod call;
 mod cell_path;
 mod expr;
 mod expression;
+mod import_pattern;
 mod operator;
 mod pipeline;
 mod statement;
@@ -12,6 +13,7 @@ pub use call::*;
 pub use cell_path::*;
 pub use expr::*;
 pub use expression::*;
+pub use import_pattern::*;
 pub use operator::*;
 pub use pipeline::*;
 pub use statement::*;
